SWITCHER-87 Terminado
