+++ src/components/Game.jsx
@@ -4,7 +4,7 @@ import '../styles/Lobby.css';
 import { GameContext } from '../contexts/GameContext.jsx';
 import Board from './Board.jsx';
 import ButtonSet from  './ButtonSet.jsx'
-
+import PlayerBox from "./PlayerBox.jsx";
 
 export default function Game({onPassTurn}) {
     const { fase, idPlayer, players, playersTurns, playersNames, idGame, setPlayers, setPlayersTurns, setPlayersNames} = useContext(GameContext);
@@ -28,42 +28,21 @@ export default function Game({onPassTurn}) {
                 {/* Cartas del jugador 1, tablero jugador 3 */}
                 <Row>
                     <Col xs="12" className="justify-content-md-center">
-                    {players.length > 1 && (() => {
-                        const otherPlayers = players.filter(player => player !== idPlayer);
-                        const firstPlayer = otherPlayers[0]; // Get the second player in the list
-                        console.log(firstPlayer)
-                        return firstPlayer ? (
-                            <h1>{playersNames[players.indexOf(firstPlayer)]}</h1>
-                        ) : null;
-                    })()}
+                        <PlayerBox boxNumber={1}/>
                     </Col>
                 </Row>
                 <Row>
                     <Col xs="auto" className="justify-content-md-left">
-                        {players.length > 2 && (() => {
-                            const otherPlayers = players.filter(player => player !== idPlayer);
-                            const secondPlayer = otherPlayers[1]; // Get the second player in the list
-                            console.log(secondPlayer);
-                            return secondPlayer ? (
-                                <h1>{playersNames[players.indexOf(secondPlayer)]}</h1>
-                            ) : null;
-                        })()}
+                        <PlayerBox boxNumber={2}/>
+                    </Col>
+                    <Col xs="auto" className="justify-content-md-right">
+                        <PlayerBox boxNumber={3}/>
                     </Col>
                     <Col xs="auto">
                     <Row>
                         <Board />
                         <ButtonSet onPassTurn={onPassTurn} />
                     </Row>
-                    </Col>
-                    <Col xs="auto" className="justify-content-md-right">
-                        {players.length > 3 && (() => {
-                            const otherPlayers = players.filter(player => player !== idPlayer);
-                            const thirdPlayer = otherPlayers[2]; // Get the second player in the list
-                            console.log(thirdPlayer);
-                            return thirdPlayer ? (
-                                <h1>{playersNames[players.indexOf(thirdPlayer)]}</h1>
-                            ) : null;
-                        })()}
                     </Col>
                 </Row>
             </Col>
